refactor: removing double code
